refactor: move core functions to utils file
