+++ src/uny-random.js
@@ -1,44 +1,20 @@
-const MANTISSA_MAX = 2 ** 23 - 1;
-const BOROSH_INIT = 1812433253;
-const TAU = Math.fround(2 * Math.PI);
-
-const toUnsigned = (num) => num >>> 0;
-const roundTo7 = (num) => Number(num.toPrecision(7));
-const borosh13 = (num) => toUnsigned(Math.imul(BOROSH_INIT, num) + 1);
-const value = (rand) => Math.fround(toUnsigned(rand & MANTISSA_MAX) / MANTISSA_MAX);
-const valueInv = (rand) => Math.fround(1 - value(rand));
-const lerp = (x, y, a) => x * (1 - a) + y * a;
-
-const rangeInt = (rand, min, max) => {
-  const result = (rand % (min - max)) + min;
-  if ( min > max) {
-    return 2 * min - result;
-  }
-  return result;
-}
-
-const rangeFloat = (rand, min, max) => {
-  const roundedMin = Math.fround(min);
-  const roundedMax = Math.fround(max);
-  return Math.fround(value(rand) * (roundedMin - roundedMax) + roundedMax);
-};
-
-const generateStateFromSeed = (seed) => {
-  const s0 = toUnsigned(seed);
-  const s1 = borosh13(s0);
-  const s2 = borosh13(s1);
-  const s3 = borosh13(s2);
-
-  return [s0, s1, s2, s3];
-};
-
-const xorshift128 = ([x, , , y]) => {
-  x ^= x << 11;
-  x ^= x >>> 8;
-  y ^= y >>> 19;
-  
-  return toUnsigned(y ^ x);
-};
+import {
+  TAU,
+  lerp,
+  roundTo7,
+  xorshift128,
+  generateStateFromSeed,
+  value,
+  valueInv,
+  rangeInt,
+  rangeFloat,
+  polarCoordinatesToVector,
+  sphericalCoordinatesToVector,
+  normalize3dVector,
+  normalizeQuaternion,
+  hopfMapping,
+  hsvToRgb,
+} from "./utils";
 
 /** Non-static implementation of the UnityEngine.Random class
  * @see {@link https://docs.unity3d.com/ScriptReference/Random.html UnityEngine.Random}
@@ -141,12 +117,14 @@ export class UnyRandom {
    * @readonly
    */
   get insideUnitCircle() {
-    const theta = valueInv(this.next) * TAU;
-    const radius = Math.sqrt(valueInv(this.next));
 
-    const x = radius * Math.cos(theta);
-    const y = radius * Math.sin(theta);
+    const polarCoordinates = {
+      theta: valueInv(this.next) * TAU,
+      radius: Math.sqrt(valueInv(this.next)),
+    }
 
+    const {x, y} = polarCoordinatesToVector(polarCoordinates);
+    
     return {
       x: roundTo7(x),
       y: roundTo7(y),
@@ -159,18 +137,13 @@ export class UnyRandom {
    */
   get insideUnitSphere() {
 
-    const phi = Math.acos(2 * valueInv(this.next) - 1);
-    const theta = valueInv(this.next) * TAU;
-    const radius = value(this.next) ** (1/3);
+    const sphericalCoordinates = {
+      phi: Math.acos(2 * valueInv(this.next) - 1),
+      theta: valueInv(this.next) * TAU,
+      radius: value(this.next) ** (1/3),
+    }
 
-    const sinTheta = Math.sin(theta);
-    const cosTheta = Math.cos(theta);
-    const sinPhi = Math.sin(phi);
-    const cosPhi = Math.cos(phi);
-
-    const x = radius * sinPhi * cosTheta;
-    const y = radius * sinPhi * sinTheta;
-    const z = radius * cosPhi;
+    const {x, y, z} = sphericalCoordinatesToVector(sphericalCoordinates);
 
     return {
       x: roundTo7(x),
@@ -184,12 +157,8 @@ export class UnyRandom {
    * @readonly
    */
   get onUnitSphere() {
-    const baseVector = this.insideUnitSphere;
-    const length = Math.sqrt(baseVector.x ** 2 + baseVector.y ** 2 + baseVector.z ** 2);
 
-    const x = baseVector.x / length;
-    const y = baseVector.y / length;
-    const z = baseVector.z / length;
+    const {x, y, z} = normalize3dVector(this.insideUnitSphere);
 
     return {
       x: roundTo7(x),
@@ -204,17 +173,14 @@ export class UnyRandom {
    */
   get rotation() {
 
-    const baseX = rangeFloat(this.next, -1, 1);
-    const baseY = rangeFloat(this.next, -1, 1);
-    const baseZ = rangeFloat(this.next, -1, 1);
-    const baseW = rangeFloat(this.next, -1, 1);
+    const baseQuaternion = {
+      x: rangeFloat(this.next, -1, 1),
+      y: rangeFloat(this.next, -1, 1),
+      z: rangeFloat(this.next, -1, 1),
+      w: rangeFloat(this.next, -1, 1),
+    };
 
-    const length = Math.sqrt(baseW ** 2 + baseX ** 2 + baseY ** 2 + baseZ ** 2) * ((baseW >= 0) || -1);
-
-    const x = baseX / length;
-    const y = baseY / length;
-    const z = baseZ / length;
-    const w = baseW / length;
+    const {x, y, z, w} = normalizeQuaternion(baseQuaternion);
 
     return {
       x: roundTo7(x),
@@ -229,22 +195,20 @@ export class UnyRandom {
    * @readonly
    */
   get rotationUniform() {
-    const u1 = value(this.next);
-    const u2 = value(this.next);
-    const u3 = value(this.next);
 
-    const x = Math.sqrt(u1) * Math.sin(TAU*u2);
-    const y = Math.sqrt(u1) * Math.cos(TAU*u2);
-    const z = Math.sqrt(1 - u1) * Math.sin(TAU*u3);
-    const w = Math.sqrt(1 - u1) * Math.cos(TAU*u3);
+    const baseVector = {
+      x: value(this.next),
+      y: value(this.next),
+      z: value(this.next),
+    }
 
-    const isWNegative = (w < 0);
+    const {x, y, z, w} = hopfMapping(baseVector);
 
     return {
-      x: roundTo7(x) * (isWNegative || -1),
-      y: roundTo7(y) * (!isWNegative || -1),
-      z: roundTo7(z) * (isWNegative || -1),
-      w: roundTo7(w) * (!isWNegative || -1),
+      x: roundTo7(x),
+      y: roundTo7(y),
+      z: roundTo7(z),
+      w: roundTo7(w),
     };
   }
 
@@ -252,29 +216,13 @@ export class UnyRandom {
    * @see {@link https://docs.unity3d.com/ScriptReference/Random.ColorHSV.html UnityEngine.Random.ColorHSV}
    */
   colorHSV(hueMin = 0, hueMax = 1, saturationMin = 0, saturationMax = 1, valueMin = 0, valueMax = 1, alphaMin = 1, alphaMax = 1) {
+
     const h = lerp(hueMin, hueMax, value(this.next));
     const s = lerp(saturationMin, saturationMax, value(this.next));
     const v = lerp(valueMin, valueMax, value(this.next));
     const a = lerp(alphaMin, alphaMax, value(this.next));
 
-    const i = Math.floor(h * 6);
-    const f = h * 6 - i;
-    const p = v * (1 - s);
-    const q = v * (1 - f * s);
-    const t = v * (1 - (1 - f) * s);
-
-    let r = 0;
-    let g = 0;
-    let b = 0;
-
-    switch (i % 6) {
-      case 0: r = v; g = t; b = p; break;
-      case 1: r = q; g = v; b = p; break;
-      case 2: r = p; g = v; b = t; break;
-      case 3: r = p; g = q; b = v; break;
-      case 4: r = t; g = p; b = v; break;
-      case 5: r = v; g = p; b = q; break;
-  }
+    const {r, g, b} = hsvToRgb(h, s, v);
 
     return {
       r: roundTo7(r),
